Update Basic Monster Knowledge Button
As the monster knowledge feature is hardcoded, it gets out of date when you retheme or rename any of the actors in foundry.

Adds a context button to the actors tab that will update the basic info (the stuff that makes up Monster Knowledge (med)) based on what is actually set in the actor.  Will also make an attempt to rename the actor in power effects for the hard description, but that is a very inexact science.

Also removes the example todo-ist code as I haven't looked at it for ages.