Tah v3 (#23)
* Refactoring to use the V3 hooks:

Groups for powers, features and inventory are pulled directly from the system.

Powers are grouped by the system

Tooltips use the new tooltip hook.

Commented out logger.debug lines in settings to try to stop the infinite loop.  Failed because it's baked into TAH, raised an issue.

* Fixed incorrect .map use that caused unresolved futures in the Power Action Builder.

This bug is also present in the v2 action generator but is not getting fixed because that code is deprecated.

Tooltip styles magically work with foxes styles.  Fox has done unknown magic, IDK.  I am taking the win.